LockStore -> LockStoreImpl
